refactor(test): move test colorLayer to own test file
